pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "segmentation_failures"
version = "2024.0.1"
authors = [
  {name = "Maximilian Zenk", email = "m.zenk@dkfz-heidelberg.de"},
]
description = "Code for the medical image segmentation failure detection benchmark"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"

dependencies = [
  "torch==2.0.1",
  "torchvision==0.15.2",
  "pytorch-lightning==2.0.1.post0",
  "numpy==1.24.2",
  "torchio==0.18.91",
  "hydra-core==1.3.2",
  "timm==0.6.13",
  "pyradiomics==3.0.1",
  "scikit-learn==1.2.2",
  "SimpleITK==2.2.1",
  "nnunetv2==2.2.1",
  "monai[nibabel]==1.3.0",
  "scikit-image",
  "loguru",
  "python-dotenv",
  "pandas",
  "rich",
  "tqdm",
  "dynamic-network-architectures",
  "tensorboard",
  "natsort",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "pre-commit",
  "matplotlib",
  "seaborn>=0.13.2",
  "flake8",
  "flake8-bugbear",
  "black",
  "pytest",
  "ipython",
  "ipykernel",
]
launcher = [
  "parallel-ssh"
]

[tool.licensecheck]
using = "PEP631"
